Add support for trigger list / enumeration
This is a pretty messy diff with too much debugging code etc. in here,
but since it is just between me and... me at this point, I'll let it
go for deployment.

This adds support for the enumeration of triggers via a list
command. This should be expanded to look at all a user's identities
and so on, but we'll get this out for people to be able to use.